requirements.txt
numpy==1.20.2
streamlit==0.82.0
tensorflow-cpu==2.5.0
pillow==8.2.0